% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadown.R
\name{listify}
\alias{listify}
\title{listify()}
\usage{
listify(tbl, axis)
}
\arguments{
\item{tbl}{a tibble}

\item{axis}{a column}
}
\value{
a list
}
\description{
listify()
}
\examples{
listify(tbl = table_depts, axis = dep) \%>\%
yaml::as.yaml()
}
